requirements.txt
click>=6.7
GitPython>=2.1
numpy>=1.13
pandas>=0.20
scikit-learn>=0.19.0
pydriller >=1.15.1
elasticsearch >=7.0.0
elasticsearch-dsl >=7.2.0
